requirements.txt
pandas==1.4.1
requests==2.27.1
python-decouple==3.6
psycopg2==2.9.3
SQLAlchemy==1.4.31
